pyproject.toml
[tool.poetry]
name = "dvc-example"
version = "0.1.0"
description = ""
authors = ["Dmitry Nesmeyanov <miradg99@gmail.com>"]
readme = "README.md"
packages = [{include = "dvc_example"}]

[tool.poetry.dependencies]
python = "^3.10"
dvc = "^3.19.0"
scikit-learn = "^1.3.0"
poetry-dotenv-plugin = "^0.2.0"
pandas = "^2.1.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
